nomad/test-requirements.txt
molecule>=3.0.0
molecule-docker
ansible-lint>=3.4.0
ansible>=2.10.5
pytest-testinfra>=6.1.0
docker
